Fixes for Hough cirsles
